requirements.txt
streamlit>=1.31.0
pypdf>=4.0.0
langchain>=0.1.0
sentence-transformers>=2.5.0
langchain-community>=0.0.20
opensearch-py>=2.4.2
faiss-cpu>=1.8.0
accelerate>=0.27.0
evaluate>=0.4.1
pdfplumber>=0.11.4
numpy>=1.26.0
pillow>=10.2.0
